Build SQL infrastructure for recreating inputs after transformations
A main goal of Chrysalis is to be able to interactively inspect tests
performed during metamorphic testing. This requires that relation chains
tested must be recorded. To save disk space, we can only save the
starting input dataset. Later, when we want to retrieve input data after
performing transformations, we can re-apply the transformations in order
to recreate that input data. This only works when the transformations
are deterministic.

While there had been some work already done to record tested relation
chains into the SQL database, this work was incomplete and new tables
were needed to be created to normalize the database.